+++ tests/testthat/test.distToNearest.R
@@ -146,7 +146,7 @@ test_that("Test cross distToNearest with model hs5f", {
     seq2 <- c("NNACG", "NACGA", "ACGAA", "CGAAC", "GAACG", "AACGT", "ACGTA", 
               "CGTAC", "GTACG", "TACGT", "ACGTN", "CGTNN")
     
-    targeting_distance <- calcTargetingDistance(HS5FModel)
+    targeting_distance <- calcTargetingDistance(HH_S5F)
     dist <- shazam:::dist5Mers(seq1, seq2, targeting_distance)
     expect_equal(dist, 1.0574, tolerance = 0.001)